Among Us!

+++ build.js
@@ -5,9 +5,8 @@ const lodash = require("lodash");
 const https = require("https");
 const fs = require("fs-extra");
 
-const prefix = 'export const token = "";\nimport { request as raw } from "https";\n' + request.toString() + "\n";
-const types = ["string", "number", "boolean"];
 const top = [];
+const types = ["string", "number", "boolean"];
 
 async function download() {
   log("Downloading docs", true);
@@ -30,9 +29,24 @@ async function download() {
 }
 
 async function parse() {
-  const api = {};
+  const endpoints = {};
+  const api = { endpoints };
 
   log("Parsing docs", true);
+  const reference = marked.lexer(await fs.readFile("docs/Reference.md", "utf-8"));
+
+  reference.filter(({ type, text }, i) => {
+    if (api.base && api.version) return;
+    if (type !== "heading") return;
+    if (text === "Base URL") {
+      const code = reference.slice(i + 1).find(({ type }) => type === "code");
+      api.base = code.text;
+    } else if (text === "API Versioning") {
+      const table = reference.slice(i + 1).find(({ type }) => type === "table");
+      api.version = +table.rows[0][0].text;
+    }
+  });
+
   const files = await fs.readdir("docs/resources");
   for await (const file of files) {
     log("Reading " + file);
@@ -120,7 +134,7 @@ async function parse() {
         });
       });
 
-      api[name] = { method, path, params, query, body };
+      endpoints[name] = { method, path, params, query, body };
     });
   }
 
@@ -129,75 +143,79 @@ async function parse() {
 
 async function compile(api) {
   log("Compiling to TypeScript", true);
-  const lib = Object.entries(api)
-    .map(([key, { method, path, params, query, body }]) => {
-      const requestArgs = ["'" + method + "'", "'" + path + "'"];
+  const prefix =
+    [
+      "const hostname = '" + new URL(api.base).hostname + "'",
+      "const prefix = '/api/v" + api.version + "'",
+      "const request = require('../request')",
+      "import { Stream } from 'stream'",
+      "interface File { name: string, value: string | Buffer | Stream }"
+    ].join("\n") + "\n";
+
+  const lib = sort(
+    Object.entries(api.endpoints).map(([key, { method, path, params, query, body }]) => {
       const args = [];
-      const fn = [];
-      method = method.toLowerCase();
+      const requestArgs = [
+        [
+          "'" + method + "'",
+          params.includes("channelId") ? "channelId" : "'none'",
+          params.includes("guildId") ? "guildId" : "'none'",
+          "'" + path + "'"
+        ].join("+':'+"),
+        "'" + method + "'",
+        "'" + path + "'"
+      ];
 
       log("Creating params for " + path);
       if (params.length)
         args.push(
           ...params.map(param => {
-            requestArgs[1] += ".replace(':" + param + "', " + param + ")";
+            const last = requestArgs.length - 1;
+            requestArgs[last] = requestArgs[last].replace(":" + param, "'+" + param + "+'").replace(/\+''$/, "");
             return param + ":string";
           })
         );
 
       log("Creating body for " + path);
       if (body) {
-        requestArgs.push("b");
         args.push("body:" + key + "Body");
-        fn.push(...parseBody(key, body));
+        requestArgs.push(...parseBody(key, body));
       } else requestArgs.push("null");
 
       log("Creating query for " + path);
       if (query) {
-        requestArgs.push("q");
         args.push("query:" + key + "Query={}");
-        fn.push(...parseQuery(key, query));
+        requestArgs.push(...parseQuery(key, query));
       } else requestArgs.push("null");
 
-      fn.push("return request(" + requestArgs + ")");
-      return "export function " + key + "(" + args + "){" + fn.join(";") + "}";
-    })
-    .sort(({ length: a }, { length: b }) => a - b)
-    .join("\n");
-
-  return prefix + top.sort(({ length: a }, { length: b }) => a - b).join("\n") + "\n" + lib;
-}
-
-function request(method, path, body, query) {
-  if (!token) Promise.reject("No token provided");
-
-  // definitely not scuffed
-  if (query) {
-    query = JSON.stringify(query);
-    if (query !== "{}") path += "?" + new URLSearchParams(JSON.parse(query)).toString();
-  }
-
-  return new Promise((resolve, reject) => {
-    const req = raw(
-      {
-        method,
-        path: "/api/v9" + path,
-        hostname: "discord.com",
-        headers: { "content-type": "application/json", authorization: "Bot " + token }
-      },
-      res => {
-        if (res.statusCode < 200 || res.statusCode >= 400) reject(res.statusCode + ": " + res.statusMessage);
-        else {
-          let text = "";
-          res.on("data", chunk => (text += chunk));
-          res.on("end", () => resolve(res.headers["content-type"] === "application/json" ? JSON.parse(text) : text));
+      let pre = key + "(" + args + "):Promise<any>{return $(" + requestArgs + ")},";
+      if (method !== "GET") {
+        let last = args.pop();
+        if (last) {
+          requestArgs.push("files");
+          if (last.startsWith("query:")) {
+            args.push("files:Array<string|File>");
+            args.push(last);
+            return pre + "\n" + key + "Files(" + args + "){return $(" + requestArgs + ")},";
+          } else {
+            args.push(last);
+            args.push("files:Array<string|File>");
+            return key + "(" + args + "):Promise<any>{return $(" + requestArgs + ")},";
+          }
         }
       }
-    );
 
-    if (body) req.write(JSON.stringify(body));
-    req.end();
-  });
+      return pre;
+    })
+  ).join("\n");
+
+  return (
+    prefix +
+    sort(top).join("\n") +
+    "\nexport = function(token: string, type: string = 'Bot') {\nconst $ = request(hostname, prefix, typeof type === 'string' ? type + ' ' + token : token)\nreturn {\n" +
+    lib +
+    "\n}\n}"
+  );
 }
 
 function validate(block) {
@@ -213,20 +231,26 @@ function boolean(string) {
   return string;
 }
 
+function sort(array) {
+  return array.sort(({ length: a }, { length: b }) => a - b);
+}
+
 function parseQuery(name, query) {
   let transform = [];
   top.push(
     "interface " +
       name +
       "Query{" +
-      Object.entries(query).map(([key, { required, type, name }]) => {
-        transform.push(name + ":query." + key);
-        return key + (required === true ? "" : "?") + ":" + (types.includes(type) ? type : "string");
-      }) +
+      sort(
+        Object.entries(query).map(([key, { required, type, name }]) => {
+          transform.push("['" + name + "',query." + key + "]");
+          return key + (required === true ? "" : "?") + ":" + (types.includes(type) ? type : "string");
+        })
+      ) +
       "}"
   );
 
-  return ["const q={" + transform + "}"];
+  return ["[" + sort(transform) + "]"];
 }
 
 function parseBody(name, body) {
@@ -235,14 +259,16 @@ function parseBody(name, body) {
     "interface " +
       name +
       "Body{" +
-      Object.entries(body).map(([key, { required, type, name }]) => {
-        transform.push(name + ":body." + key);
-        return key + (required === true ? "" : "?") + ":" + (types.includes(type) ? type : "any");
-      }) +
+      sort(
+        Object.entries(body).map(([key, { required, type, name }]) => {
+          transform.push(name + ":body." + key);
+          return key + (required === true ? "" : "?") + ":" + (types.includes(type) ? type : "any");
+        })
+      ) +
       "}"
   );
 
-  return ["const b={" + transform + "}"];
+  return ["{" + sort(transform) + "}"];
 }
 
 let loggingLevel = "none";
@@ -253,21 +279,42 @@ function log(string, important) {
 }
 
 (async () => {
-  if (process.argv[2]) loggingLevel = process.argv[2];
+  const bin = process.cwd() + "/node_modules/.bin/";
+  const package = require("./package.json");
+  const options = process.argv
+    .slice(2)
+    .filter(a => a.startsWith("-"))
+    .map(a => a.replace(/-+/, "").toLowerCase());
 
-  const exists = await fs.pathExists("docs");
-  if (!exists) await download();
-  const api = await parse();
-  if (!process.argv.slice(2).some(a => a.toLowerCase() === "-k" || a.toLowerCase() === "--keep"))
+  if (options.includes("a") || options.includes("all")) options.push(..."pdgt".split(""));
+  if (process.argv[2] && !process.argv[2].startsWith("-")) loggingLevel = process.argv[2];
+  if (!(await fs.pathExists("docs"))) await download();
+  const parsed = await parse();
+
+  if (options.includes("p") || options.includes("parsed"))
+    await fs.writeFile("parsed.json", JSON.stringify(parsed, null, 2));
+  else if (await fs.pathExists("parsed.json")) await fs.unlink("parsed.json");
+  if (!options.includes("d") && !options.includes("docs") && !options.includes("documentation"))
     await fs.rm("docs", { recursive: true });
 
-  const compiled = await compile(api);
+  const compiled = await compile(parsed);
   await fs.writeFile("index.ts", compiled);
 
+  if (options.includes("g") || options.includes("gen") || options.includes("generate")) {
+    log("Building Docs", true);
+    await fs.writeFile("tsconfig.json", JSON.stringify({}));
+    await execSync(
+      bin +
+        'typedoc index.ts --out web --theme default --readme README.MD --name "' +
+        package.name +
+        '" --includeVersion'
+    );
+    await fs.unlink("tsconfig.json");
+  } else if (await fs.pathExists("web")) await fs.rm("web", { recursive: true });
+
   log("Compiling to JavaScript", true);
-  execSync(process.cwd() + "/node_modules/.bin/tsc index.ts -d --outdir dist");
-  if (!process.argv.slice(2).some(a => a.toLowerCase() === "-k" || a.toLowerCase() === "--keep"))
-    await fs.unlink("index.ts");
+  execSync(bin + "tsc index.ts -d --outdir dist");
+  if (!options.includes("t") && !options.includes("ts") && !options.includes("typescript")) await fs.unlink("index.ts");
 
   log("Completed", true);
 })();
